+++ lib/presentation/widgets/attachment/mobile_attachment.dart
@@ -1,0 +1,18 @@
+// mobile_boarding.dart
+import 'package:flutter/material.dart';
+
+
+class MobileAttachment extends StatelessWidget {
+  const MobileAttachment({super.key});
+
+
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+
+      body: Container(
+        child: Text('Mobile Attachment Screen'),
+      ));
+  }
+}