+++ lib/presentation/widgets/boarding/desktop_boarding.dart
@@ -3,10 +3,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:lazycopy/core/utils/my_background.dart';
+import 'package:lazycopy/core/utils/my_nagivation.dart';
 import 'package:lazycopy/presentation/components/boarding/box_explaning.dart';
 import 'package:lazycopy/presentation/components/boarding/first_phrase.dart';
 import 'package:lazycopy/presentation/components/boarding/middle_button.dart';
 import 'package:lazycopy/presentation/components/boarding/second_phrase.dart';
+import 'package:lazycopy/presentation/widgets/attachment/desktop_attachment.dart';
 
 
 class DesktopBoarding extends StatefulWidget {
@@ -69,7 +71,7 @@ class _DesktopBoardingState extends State<DesktopBoarding> {
                   minimumSize: const Size(200, 80),
                   iconSize: 30,
                   onPressed: () {
-                    print("Button pressed!");
+                    NavigationService.push(DesktopAttachment());
                   },
                 ),
                   ),